requirements.txt
bidict==0.23.1
distlib==0.3.8
EasyProcess==1.1
entrypoint2==1.1
h11==0.16.0
MouseInfo==0.1.3
mss==9.0.1
paho-mqtt==2.1.0
Pillow==10.0.0
PyAutoGUI==0.9.54
pycodestyle==2.12.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytweening==1.0.7
simple-websocket==1.1.0
striprtf==0.0.26
wsproto==1.2.0
